Don't fire onBlur with synthetic event in handlChange
A lot of libraries (most notably, redux-form) expect the blur event to
have an event.target.value defined, like it would for the blur event of
a native input element.

I could have munged the synthetic event to mock up these values, but I
didn't quite see the need; it's not clear to me that a value change
semantically justifies a call to onBlur. If someone has a use case where
this *is* required, I'm open to changing this.